Support much larger input sizes with u64 instead of u32

+++ src/main.rs
@@ -22,12 +22,12 @@ struct Args {
 
     /// Number of seats to apportion [default: 435]
     #[clap(short, long)]
-    seats: Option<u32>,
+    seats: Option<u64>,
 
     /// Multiplier to use with the cube root rule (e.g. specify 2 to get a total number of seats
     /// equal to twice the cube root of the total population)
     #[clap(short, value_name = "MULTIPLIER")]
-    cube_root_multiplier: Option<u32>,
+    cube_root_multiplier: Option<u64>,
 
     /// Whether to use the cube root rule (with a multiplier of 1) to determine the total number of seats
     #[clap(short = 'C')]
@@ -35,7 +35,7 @@ struct Args {
 
     /// Number of seats to apportion to the smallest state using the Wyoming rule
     #[clap(short, value_name = "SEATS")]
-    wyoming_seats: Option<u32>,
+    wyoming_seats: Option<u64>,
 
     /// Whether to use the Wyoming rule with 1 seat for the smallest state
     #[clap(short = 'W')]
@@ -49,8 +49,8 @@ fn main() {
 
     // Initialize priority queue based on input file and find total population
     let mut queue = BinaryHeap::new();
-    let mut total_population = 0u32;
-    let mut min_population = u32::MAX;
+    let mut total_population = 0u64;
+    let mut min_population = u64::MAX;
     for line in fs::read_to_string(input_file_path).expect("Something went wrong reading the input file.").lines() {
         let pair = line.split("\t").collect::<Vec<&str>>();
         let name: String = pair[0].parse().unwrap();
@@ -68,13 +68,13 @@ fn main() {
     let seats = if let Some(seats) = args.seats {
         seats
     } else if let Some(seats) = args.wyoming_seats {
-        ((total_population as f64 / min_population as f64) * seats as f64).round() as u32
+        ((total_population as f64 / min_population as f64) * seats as f64).round() as u64
     } else if args.wyoming_rule {
-        (total_population as f64 / min_population as f64).round() as u32
+        (total_population as f64 / min_population as f64).round() as u64
     } else if let Some(multiplier) = args.cube_root_multiplier {
-        ((total_population as f64).powf(1f64 / 3f64) * multiplier as f64).round() as u32
+        ((total_population as f64).powf(1f64 / 3f64) * multiplier as f64).round() as u64
     } else if args.cube_root_rule {
-        (total_population as f64).powf(1f64 / 3f64).round() as u32
+        (total_population as f64).powf(1f64 / 3f64).round() as u64
     } else {
         435
     };
@@ -82,10 +82,10 @@ fn main() {
 
     // Handle case where the number of seats is less than the number of states
     let mut seats_remaining = seats;
-    if queue.len() as u32 > seats_remaining {
+    if queue.len() as u64 > seats_remaining {
         panic!("You must specify at least {} seats so that each state gets at least one seat.", queue.len());
     }
-    seats_remaining -= queue.len() as u32;
+    seats_remaining -= queue.len() as u64;
 
     // Apportion all remaining seats using the priority queue
     while seats_remaining > 0 {
